Added bool to stay on same height when climbing up
Bool IsRopeClimbingActive was added to BP_VRCharacter. Toggling this bool on, makes it so that the player can climb up themselves, and stay on that height too. It also deactives the pulling up feature. If put on false, pulling up is reactivated, and climbing up yourself makes you slide down