#101 Added belt movement behaviour + fixes
Belt behaviour is now finished. In that regard, one bug exists, namely the item not resetting sometimes on very specific circumstances. Another issue is also present, where the engine crashes, when trying to grip multiple items. Although it's a seperate issue from this implementation, it is essential to be fixed, in order for the belt to work completely well.