#104 Adaptive haptic feedback implemented
Instrument with adaptive haptic feedback was added. Based on the collision depth measured, it will scale the haptics from 0 to 1, also taking into account, at which place the hands are grabbing onto the instrument. The character takes charge of applying the haptic feedback, while the instrument calculates and offers functions, to determine the scale and check for collisions.