#113 #115 Added gravity function to character & bugfix
BP_VRCharacter now has a function called "SetGravity", which is able to toggle whether the player is able to be stuck in the air or not. Though it is not in use yet, it is ready to be implemented in different levels that need this, when it comes to testing for example. Catapulting effect is presumably fixed, at least it was not able to reproduced. Tests will show whether it holds or not.