#118 Fixed bristle blaster belt issue & start #120
Removing the scene root component, and replacing it with the actual static mesh fixed the belt issue, where it was unable to set the position of the actor. A component was created, which will enable being able to paint from any other BP more easily, by using the provided functions.