#120 Added drawing component
Drawing component was added, to be more easily able to draw textures onto a material from an actor.  Outline component was also tested together with painting, and worked fine. There are some issues with painting though, which are apparent when testing as FPS_VRCharacter. If this issues of wrong disabled collisions is still active in normal VR, it will be fixed as next task.