#132 Rope is climbable but needs polishing
Climbing the rope works now. By implementing custom functions, which utilize the previous climbing logic from VRE, logic was able to be called at the appropriate places.

At the moment, when climbing, the rope still stutters. This will be subject to fix in the next push. Additionally, the hand sometimes still holds onto the rope, even though it was let go of. This is most likely tied to the processing of the climb mode.

One setting that was also changed in the project settings, are substeps. This allows for physics ticks to have more time in between, to handle collisions. This aids the rope, as it struggles to keep up with all interactions, when length is too long.