#132 #133 #135 Rope + Outline + Patch Painting
The rope is now climbable with physics activated. It's now velocity based, not using the VRE provided climb-movement functionality. The Z axis was also constrained for the rope as well, to not interfere with climbing anymore.

For outline, a material using post processing was added. With this, anything is able to be outlined in runtime, if the outline component is attached, and a post processing volume exists in the level.

For mask painting, logic was added to support lerping values such as metallic and roughness. One raycast was also removed, which makes the paint check work better and more efficient. The textures for the wind turbine patch need to be replaced and improved though, as they need to look differently and also need to be upscaled.