Replaced BP with new Bristle Blaster SkeletonMesh
A new blueprint was added, which uses a skeleton mesh instead of a static mesh for the bristle blaster (with the actual fbx). It holds the same code as the old one, only changing the type to Skeleton Mesh instead. Logic has additionally been changed to fit this new blueprint, which results in the old BP, to not be usable anymore. However, it will still be kept for now, as a potential reference point, if issues should arise. The model that was placed inside, will be replaced at some later point, as textures need to be applied still (which will result in the mesh being reworked slightly).