Added support for multiple canvases (PaintComp)
A bug was causing the material to disappear, when multiple paint components were active inside a single level. Now it supports checking the states of multiple canvases, appropriately handling dynamic material instances, and saving values seperately inside arrays for each canvas. This has a downside though, as now every paint component is handling the canvases seperately on its own. This will not be fixed though, as it's only causing issues in edge cases, and leaving it alone will save a significant amount of time.